Rename newsletter subscription model and controllers
