--- src/models/Newsletter.js
@@ -1,12 +1,0 @@
-const mongoose = require('mongoose');
-
-const NewsletterSchema = new mongoose.Schema(
-  {
-    email: String
-  },
-  {
-    timestamps: true
-  }
-);
-
-module.exports = mongoose.model('Newsletter', NewsletterSchema);
