+++ src/models/NewsletterSubscription.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose');
+
+const NewsletterSubscriptionSchema = new mongoose.Schema(
+  {
+    email: String
+  },
+  {
+    timestamps: true
+  }
+);
+
+module.exports = mongoose.model(
+  'newsletter_subscriptions',
+  NewsletterSubscriptionSchema
+);
